workaround bulk delete limitation

+++ underfs/s3a/src/main/java/alluxio/underfs/s3a/S3AUnderFileSystem.java
@@ -309,16 +309,24 @@ public class S3AUnderFileSystem extends ObjectUnderFileSystem {
   protected List<String> deleteObjects(List<String> keys) throws IOException {
     Preconditions.checkArgument(keys != null && keys.size() <= getListingChunkLengthMax());
     try {
-      List<DeleteObjectsRequest.KeyVersion> keysToDelete = new ArrayList<>();
-      for (String key : keys) {
-        keysToDelete.add(new DeleteObjectsRequest.KeyVersion(key));
-      }
-      DeleteObjectsResult deletedObjectsResult =
-          mClient.deleteObjects(new DeleteObjectsRequest(mBucketName).withKeys(keysToDelete));
       List<String> deletedObjects = new ArrayList<>();
-      for (DeleteObjectsResult.DeletedObject deletedObject : deletedObjectsResult
-          .getDeletedObjects()) {
-        deletedObjects.add(deletedObject.getKey());
+      if (Configuration.getBoolean(PropertyKey.UNDERFS_S3A_BULK_DELETE_ENABLED)) {
+        List<DeleteObjectsRequest.KeyVersion> keysToDelete = new ArrayList<>();
+        for (String key : keys) {
+          keysToDelete.add(new DeleteObjectsRequest.KeyVersion(key));
+        }
+        DeleteObjectsResult deletedObjectsResult =
+            mClient.deleteObjects(new DeleteObjectsRequest(mBucketName).withKeys(keysToDelete));
+        for (DeleteObjectsResult.DeletedObject deletedObject : deletedObjectsResult
+            .getDeletedObjects()) {
+          deletedObjects.add(deletedObject.getKey());
+        }
+      } else {
+        for (String key : keys) {
+          if (deleteObject(key)) {
+            deletedObjects.add(key);
+          }
+        }
       }
       return deletedObjects;
     } catch (AmazonClientException e) {
